Separation des Controller et Service Tour & Show
