+++ src/Api/MaViCS/MaViCS.Business/Services/TourService.cs
@@ -9,12 +9,10 @@ namespace MaViCS.Business.Services
     public class TourService : ITourService
     {
         public readonly ITourRepository _tourRepository;
-        public readonly IShowRepository _showRepository;
 
-        public TourService(ITourRepository tourRepository, IShowRepository showRepository)
+        public TourService(ITourRepository tourRepository)
         {
             _tourRepository = tourRepository;
-            _showRepository = showRepository;
         }
 
         public async Task<IEnumerable<TourDto>> GetTours()
@@ -29,13 +27,6 @@ namespace MaViCS.Business.Services
             var tours = await _tourRepository.GetToursByTalent(talentId);
 
             return tours.Select(x => x.ToTourDto()).ToList();
-        }
-
-        public async Task<IOrderedEnumerable<ShowDto>> GetShowsByTour(long tourId)
-        {
-            var shows = await _showRepository.GetShowsByTour(tourId);
-
-            return shows.Select(x => x.ToShowDto()).OrderBy(x => x.Date);
         }
 
         public async Task<TourDto?> GetTourById(long id)
@@ -67,33 +58,6 @@ namespace MaViCS.Business.Services
             return tour?.ToTourDto();
         }
 
-        public async Task<ShowDto?> AddShow(long tourId, CreateOrUpdateShowDto showDto)
-        {
-            var tour = await _tourRepository.GetTourById(tourId);
-
-            if (tour == null)
-                return null;
-
-            var show = showDto.ToShow();
-            show.TourId = tourId;
-
-            show = await _showRepository.AddShow(show);
-            return show?.ToShowDto();
-        }
-
-        public async Task<ShowDto?> UpdateShow(long showId, CreateOrUpdateShowDto showDto)
-        {
-            var show = await _showRepository.GetShowById(showId);
-
-            if (show == null)
-                return null;
-
-            show = show.UpdateShow(showDto);
-
-            show = await _showRepository.AddShow(show);
-            return show?.ToShowDto();
-        }
-
         public async Task<bool> ArchiveTour(long id)
         {
             return await _tourRepository.ArchiveTour(id);
@@ -102,11 +66,6 @@ namespace MaViCS.Business.Services
         public async Task<bool> DeleteTour(long id)
         {
             return await _tourRepository.DeleteTour(id);
-        }
-
-        public async Task<bool> DeleteShow(long id)
-        {
-            return await _showRepository.DeleteShow(id);
         }
 
     }
