+++ src/Api/MaViCS/MaViCS.sln
@@ -3,12 +3,40 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MaViCS", "MaViCS\MaViCS.csproj", "{6C9342BA-FF9E-491A-8F7C-35F35EF9475B}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MaViCS.Business", "MaViCS.Business\MaViCS.Business.csproj", "{1B96F9EA-5CCD-4416-9548-AA94C32A5B53}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MaViCS.Domain", "MaViCS.Domain\MaViCS.Domain.csproj", "{DCD2CE9D-AF7A-48C8-A5A0-DE303616AD77}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{85647665-4318-48D1-A950-66041D0BE241}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6C9342BA-FF9E-491A-8F7C-35F35EF9475B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6C9342BA-FF9E-491A-8F7C-35F35EF9475B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6C9342BA-FF9E-491A-8F7C-35F35EF9475B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6C9342BA-FF9E-491A-8F7C-35F35EF9475B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1B96F9EA-5CCD-4416-9548-AA94C32A5B53}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1B96F9EA-5CCD-4416-9548-AA94C32A5B53}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1B96F9EA-5CCD-4416-9548-AA94C32A5B53}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1B96F9EA-5CCD-4416-9548-AA94C32A5B53}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DCD2CE9D-AF7A-48C8-A5A0-DE303616AD77}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DCD2CE9D-AF7A-48C8-A5A0-DE303616AD77}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DCD2CE9D-AF7A-48C8-A5A0-DE303616AD77}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DCD2CE9D-AF7A-48C8-A5A0-DE303616AD77}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {490EB860-CA54-4D13-800D-9C7A13EC816A}
 	EndGlobalSection
 EndGlobal
